Updates to the Robot Code and Methods
Switched from using Solenoid to DoubleSolenoid as I remembered I used it last year. Also I have removed the NotDefiendException as it was just for some testing and shouldn't be in the code base as it would not help the robot.

AutoDrive has been Converted to use the OOP method unlike last year's robot code where it called the Static Instance of the `DifferentialDrive`.

Co-Authored-By: ashman9704 <ashman9704@users.noreply.github.com>